Implemented Part 5: exception handling, mappers, logging filter

+++ src/main/java/com/backend/resource/SensorRoom.java
@@ -2,6 +2,8 @@ package com.backend.resource;
 
 import com.backend.dao.*;
 import com.backend.model.Room;
+import com.backend.model.Sensor;
+import com.backend.exception.*;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import java.util.List;
@@ -9,6 +11,7 @@ import java.util.List;
 @Path("/rooms")
 public class SensorRoom {
     
+    private GenericDAO<Sensor> sensorDAO = new GenericDAO<>(MockDatabase.SENSOR);
     private GenericDAO<Room> sensorRoomDAO = new GenericDAO<>(MockDatabase.ROOM);
     
     @GET
@@ -33,6 +36,20 @@ public class SensorRoom {
     @DELETE
     @Path("/{roomId}")
     public void deleteRoom(@PathParam("roomId") String roomId) {
+        
+        boolean hasSensor = false;
+       
+        for (Sensor sensor : sensorDAO.getAll()) {
+            if(roomId.equals(sensor.getRoomId())){
+                hasSensor = true;
+                break;
+            }
+        }
+        
+        if (hasSensor){
+            throw new RoomNotEmptyException("This room cannot be deleted as it is currently occupied by active sensors");
+        }
+        
         sensorRoomDAO.delete(roomId);
     }
